requirements.txt
google-genai>=1.0.0
datasets>=2.14.0,<3.0.0
numpy>=1.24.0
tqdm>=4.65.0
